requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.8.1
pydantic>=2.6.1
SQLAlchemy>=2.0.25
asyncpg>=0.29.0
python-jose>=3.3.0 